for real for real

+++ TestApplication/app/src/main/java/com/example/testapplication/ArithmeticsStats.kt
@@ -34,7 +34,7 @@ class ArithmeticsStats : Fragment() {
             return
         }
         binding.gamesPlayedText.text =viewmodel.ArithmeticsGamesPlayed.toString()
-        var winP = (viewmodel.ArithmeticsWinPercentage * 100).toInt().toString()
+        var winP = (viewmodel.ArithmeticsWinPercentage * 100).toString()
         if(winP.length > 6){ binding.winPercentageText.text = winP.substring(0,7)+"%" }
         else{ binding.winPercentageText.text = winP+"%" }
