+++ TestApplication/app/src/main/java/com/example/testapplication/KompassStats.kt
@@ -37,7 +37,7 @@ class KompassStats : Fragment() {
             return
         }
         binding.gamesPlayedText.text =viewmodel.CompassGamesPlayed.toString()
-        var winP = (viewmodel.CompassWinPercentage * 100).toInt().toString()
+        var winP = (viewmodel.CompassWinPercentage * 100).toString()
         if(winP.length > 6){ binding.winPercentageText.text = winP.substring(0,7)+"%" }
         else{ binding.winPercentageText.text = winP+"%" }
